Refactor mint-diet into route modules
