+++ mock-platform/mocks/mint-diet/src/components/index.ts
@@ -1,0 +1,16 @@
+export { Layout } from "./Layout";
+export {
+  DayNav,
+  EntryForm,
+  MealSlotCard,
+  SearchResultRow,
+  SummaryPanel,
+} from "./log";
+export {
+  IngredientTable,
+  PlanCard,
+  PlanDayGrid,
+  PlanDetailPage,
+  PlanForm,
+  SlotEditorPage,
+} from "./plans";
